rename to highlight to match config in file
